requirements.txt
flask==3.0.0
python-owasp-zap-v2.4==0.0.20
pdfkit==1.0.0  